Creating global.db upgrade schema

+++ src/init/wazuh/wazuh.sh
@@ -98,8 +98,12 @@ WazuhUpgrade()
     fi
 
     # Remove existing SQLite databases
-
-    rm -f $DIRECTORY/var/db/global.db*
+    if [-f "$DIRECTORY/var/db/global.db"]; then
+        cp $DIRECTORY/var/db/global.db* $DIRECTORY/queue/db/global.db*
+        chmod 640 $DIRECTORY/queue/db/global.db*
+        chown ossec:ossec $DIRECTORY/queue/db/global.db*
+        rm -f $DIRECTORY/var/db/global.db*
+    fi
     rm -f $DIRECTORY/var/db/.profile.db*
     rm -f $DIRECTORY/var/db/.template.db*
     rm -f $DIRECTORY/var/db/agents/*
